+++ app/mailers/user_mailer.rb
@@ -1,2 +1,10 @@
 class UserMailer < ApplicationMailer
+
+  def contact(name, email, message)
+    subject = "#{name} - Website Message"
+    @email_to = 'reels@conductorproductions.com'
+    @message = message
+    mail(to: @email_to, subject: subject, reply_to: email)
+  end
+
 end